[ADD] Ajout externalisation des controles
